PP-11212 Rename Stripe files and routes
We had the routes `/stripe/basic/` and files in a folder named "basic"
for creating Stripe accounts.

This is because previously there was a "not basic" way of creating
accounts where we needed to configure more account details. Since the
old routes and files no longer exist, "basic" no longer makes any
sense. So rename the files and routes accordingly.
