+++ src/web/modules/stripe/confirm-create-test-account.njk
@@ -28,7 +28,7 @@
     }) }}
   {% endif %}
 
-  <form method="POST" action="/stripe/basic/create-test-account">
+  <form method="POST" action="/stripe/create-test-account">
     <p class="govuk-body">You are about to create a Stripe account for the service. This will create:</p>
     <ul class="govuk-list govuk-list--number">
       <li>Stripe test connect account through the Stripe API</li>
